Removed Fyfe article ref
KIB indicates that the pdf of the Fyfe article was never intended to be part of this entry. (KDH probably added it later.) In "Copy of the Artifact," removed mention of Fyfe article and replaced with "forthcoming" presumably to be a copy of Cordell's forking the article rather than the original article.